% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_surv_est.R
\name{conditional_surv_est}
\alias{conditional_surv_est}
\title{Estimate conditional survival with a 95\% confidence interval}
\usage{
conditional_surv_est(basekm, t1, t2)
}
\arguments{
\item{basekm}{\code{survfit} object}

\item{t1}{the time on which to condition}

\item{t2}{the survival time to estimate}
}
\value{
A list where \code{cs_est} is the conditional survival estimate,
\code{cs_lci} is the lower bound of the 95% confidence interval and
\code{cs_uci} is the upper bound of the 95% confidence interval
}
\description{
\code{conditional_surv_est} estimates the Kaplan-Meier conditional survival at
fixed time points and produces a 95\% confidence interval
}
\details{
For example, if \code{t1} = 2 and \code{t2} = 5, the function
will return the probability of surviving to year 5 conditioned on having
already survived to year 2. See the vignette
at \href{http://www.emilyzabor.com/condsurv/articles/estimate_cs.html}{http://www.emilyzabor.com/condsurv/articles/estimate_cs.html} for
details on calculations of conditional survival estimates and confidence
intervals, and examples.
}
