% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condKMapp.R
\name{condKMapp}
\alias{condKMapp}
\title{Interactive conditional survival Shiny app}
\usage{
condKMapp(.basekm)
}
\arguments{
\item{.basekm}{\code{survfit} object}
}
\value{
The output includes 1) a plot with one overall Kaplan-Meier survival
curve and an additional Kaplan-Meier survival curve for each time on which
you are conditioning and 2) a table with conditional survival estimates at
the specified survival time along with their associated confidence intervals,
with one row for each time on which you are conditioning.
}
\description{
\code{condKMapp} produces an interactive Shiny app containing output from
both the \code{condKMest} function and the \code{condKMplot} function.
}

