% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condKMplot.R
\name{condKMplot}
\alias{condKMplot}
\title{Generate conditional survival plots}
\usage{
condKMplot(.basekm, .at, .main = NULL, .xlab = NULL,
  .ylab = "Survival probability", .lwd = 1, .mark = F)
}
\arguments{
\item{.basekm}{\code{survfit} object}

\item{.at}{vector of times on which to condition}

\item{.main}{plot title}

\item{.xlab}{x-axis label}

\item{.ylab}{y-axis label, defaults to "Survival probability"}

\item{.lwd}{plot line width, defaults to 1}

\item{.mark}{controls whether censoring times are marked on curves,
defaults to F}
}
\value{
A plot with a line for the overall Kaplan-Meier plot and one
additional line for each value in \code{.at}
}
\description{
\code{condKMplot} produces a Kaplan-Meier plot for a variety of times on
which to condition
}
