% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKM.R
\name{PKM}
\alias{PKM}
\title{Presmoothed Kaplan-Meier product-limit estimate of survival.}
\usage{
PKM(time, status, t)
}
\arguments{
\item{time}{Survival time of the process.}

\item{status}{Censoring indicator of the survival time of the process; 0 if
the survival time is censored and 1 otherwise.}

\item{t}{The time for obtaining survival estimates.}
}
\description{
This function provides survival estimates using the presmoothed
product-limit Kaplan-Meier estimator.
}
\examples{

obj <- with(colonCS, survCS(time1, event1, Stime, event))
PKM(time = obj$Stime, status = obj$event, t = 1095)

}
\references{
R. Cao, I. Lopez-de Ullibarri, P. Janssen, and N. Veraverbeke.
Presmoothed kaplan-meier and nelsonaalen estimators. Journal of
Nonparametric Statistics, 17:31-56, 2005.

G. Dikta. On semiparametric random censorship models. Journal of Statistical
Planning and Inference, 66:253-279, 1998.

E. Kaplan and P. Meier. Nonparametric estimation from incomplete
observations. Journal of the American Statistical Association, 53:457-481,
1958.
}
\seealso{
\code{\link{KM}}
}
\author{
Luis Meira-Machado and Marta Sestelo
}
