\name{condSURV-package}
\alias{condSURV-package}
\alias{condSURV}
\docType{package}
\title{condSURV: \packageTitle{condSURV}
%A package for nonparametric estimation of the survival functions for ordered multivariate failure time data.
}
\description{
Newly developed methods for the estimation of
the conditional survival function are implemented in this package. The condSURV package implements nonparametric and
semiparametric estimators for these quantities. The package also implements feasible
estimation methods for these quantities conditionally on current or past covariate measures. Other
related estimators are also implemented in the package. One of these estimators is the Kaplan-Meier
estimator typically assumed to estimate the survival function. A modification of the Kaplan-Meier
estimator, based on a preliminary estimation (presmoothing) of the
censoring probability for the survival time, given the available information is also implemented.

}


\author{
Luis Meira-Machado and Marta Sestelo.

Maintainer: Luis Meira-Machado, lmachado@math.uminho.pt
}

\references{
L. Meira-Machado, M. Sestelo, and A. Goncalves (2016). Nonparametric estimation of the survival function
for ordered multivariate failure time data: a comparative study. Biometrical Journal, 58(3), 623--634.
}


