% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_conditional_surv.R
\name{gg_conditional_surv}
\alias{gg_conditional_surv}
\title{Generate conditional survival plots using ggplot2}
\usage{
gg_conditional_surv(basekm, at, main = NULL, xlab = "Years",
  ylab = "Survival probability", lwd = 1)
}
\arguments{
\item{basekm}{\code{survfit} object}

\item{at}{vector of times on which to condition}

\item{main}{plot title}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label, defaults to "Survival probability"}

\item{lwd}{plot line width, defaults to 1}
}
\value{
A ggplot with a line for the overall Kaplan-Meier plot and one
additional line for each value in \code{at}

#' @details See the vignette
at \href{http://www.emilyzabor.com/condsurv/articles/plot_cs.html}{http://www.emilyzabor.com/condsurv/articles/plot_cs.html} for
details and examples.
}
\description{
\code{gg_conditional_surv} produces a Kaplan-Meier plot for a variety of times on
which to condition using ggplot2
}
